--- a/agent/Cargo.toml
+++ b/agent/Cargo.toml
@@ -113,6 +113,10 @@ tonic-build = "0.8.0"
 [features]
 
 [[bench]]
+name = "common"
+harness = false
+
+[[bench]]
 name = "flow_generator"
 harness = false