--- a/agent/src/common/tap_types.rs
+++ b/agent/src/common/tap_types.rs
@@ -19,7 +19,10 @@ use std::{
     fmt,
     net::Ipv4Addr,
     str::FromStr,
-    sync::{Arc, Mutex},
+    sync::{
+        atomic::{AtomicU16, Ordering},
+        Mutex,
+    },
 };
 
 use log::warn;
@@ -31,18 +34,19 @@ use super::{enums::TapType, XflowKey};
 const VLAN_MAX: u16 = 4096;
 
 pub struct TapTyper {
-    packet: Arc<Mutex<[TapType; (VLAN_MAX + 1) as usize]>>,
-    xflow: Arc<Mutex<HashMap<XflowKey, TapType>>>,
+    packet: [AtomicU16; (VLAN_MAX + 1) as usize],
+    xflow: Mutex<HashMap<XflowKey, TapType>>,
     //xflowmissed 没有删除操作，只有插入操作，这是业务要求(仅打印一次)，问过苑超说key不会一直增长，应该不会有内存泄漏问题
-    _xflow_missed: Arc<Mutex<HashSet<XflowKey>>>,
+    _xflow_missed: Mutex<HashSet<XflowKey>>,
 }
 
 impl TapTyper {
+    const TAP_TYPE_ANY: AtomicU16 = AtomicU16::new(0);
     pub fn new() -> Self {
         Self {
-            packet: Arc::new(Mutex::new([TapType::Any; (VLAN_MAX + 1) as usize])),
-            xflow: Arc::new(Mutex::new(HashMap::new())),
-            _xflow_missed: Arc::new(Mutex::new(HashSet::new())),
+            packet: [Self::TAP_TYPE_ANY; (VLAN_MAX + 1) as usize],
+            xflow: Mutex::new(HashMap::new()),
+            _xflow_missed: Mutex::new(HashSet::new()),
         }
     }
 
@@ -51,16 +55,15 @@ impl TapTyper {
             return None;
         }
 
-        let mut packet_guard = self.packet.lock().unwrap();
-        let vlan = vlan as usize;
-
-        if packet_guard[vlan] == TapType::Any {
+        let p = &self.packet[vlan as usize];
+        let tt = p.load(Ordering::Relaxed).try_into().unwrap();
+        if tt == TapType::Any {
             warn!("vlan {}'s tap_type is unknown", vlan);
-            packet_guard[vlan] = TapType::Unknown;
-            return None;
+            p.store(TapType::Unknown.into(), Ordering::Relaxed);
+            None
+        } else {
+            Some(tt)
         }
-
-        Some(packet_guard[vlan])
     }
 
     pub fn get_tap_type_by_xflow_key(&self, xflow_key: &XflowKey) -> Option<TapType> {
@@ -79,7 +82,9 @@ impl TapTyper {
     }
 
     pub fn on_tap_types_change(&self, tap_types: Vec<trident::TapType>) {
-        let mut packet = [TapType::Any; (VLAN_MAX + 1) as usize];
+        for tap in self.packet.iter() {
+            tap.store(0, Ordering::Relaxed);
+        }
         let mut xflow = HashMap::new();
         for tap_type in tap_types {
             match tap_type.packet_type() {
@@ -91,13 +96,10 @@ impl TapTyper {
                         warn!("invalid vlan({}) or tap_type is {}", vlan, tap);
                         continue;
                     }
-                    match TapType::try_from(tap) {
-                        Ok(tap_type) => {
-                            packet[vlan as usize] = tap_type;
-                        }
-                        Err(err) => {
-                            warn!("prase tap_type from protocol buffer TapType error: {}", err);
-                        }
+                    if let Err(e) = TapType::try_from(tap) {
+                        warn!("parse tap_type from protocol buffer TapType error: {}", e);
+                    } else {
+                        self.packet[vlan as usize].store(tap, Ordering::Relaxed);
                     }
                 }
                 _ => {
@@ -129,7 +131,6 @@ impl TapTyper {
             }
         }
 
-        *self.packet.lock().unwrap() = packet;
         *self.xflow.lock().unwrap() = xflow;
     }
 }
@@ -144,13 +145,12 @@ impl fmt::Display for TapTyper {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let packet: Vec<(u16, TapType)> = self
             .packet
-            .lock()
-            .unwrap()
             .iter()
             .enumerate()
             .filter_map(|(vlan, tap_type)| {
-                if *tap_type != TapType::Any {
-                    Some((vlan as u16, *tap_type))
+                let tap_type = TapType::try_from(tap_type.load(Ordering::Relaxed)).unwrap();
+                if tap_type != TapType::Any {
+                    Some((vlan as u16, tap_type))
                 } else {
                     None
                 }