--- a/agent/src/flow_generator/flow_map.rs
+++ b/agent/src/flow_generator/flow_map.rs
@@ -1024,9 +1024,7 @@ impl FlowMap {
             self.update_endpoint_and_policy_data(node, meta_packet);
         } else {
             // copy endpoint and policy data
-            meta_packet.policy_data.replace(Arc::new(
-                node.policy_data_cache[meta_packet.lookup_key.direction as usize].clone(),
-            ));
+            meta_packet.policy_data.replace(node.policy_data_cache[meta_packet.lookup_key.direction as usize].clone());
             match meta_packet.lookup_key.direction {
                 PacketDirection::ClientToServer => {
                     meta_packet
@@ -1747,7 +1745,7 @@ impl FlowMap {
 
         // update policy data
         if meta_packet.policy_data.is_some() {
-            node.policy_data_cache[meta_packet.lookup_key.direction as usize] = PolicyData {
+            node.policy_data_cache[meta_packet.lookup_key.direction as usize] = Arc::new(PolicyData {
                 acl_id: meta_packet.policy_data.as_ref().unwrap().acl_id,
                 npb_actions: meta_packet
                     .policy_data
@@ -1756,7 +1754,7 @@ impl FlowMap {
                     .npb_actions
                     .clone(),
                 action_flags: meta_packet.policy_data.as_ref().unwrap().action_flags,
-            };
+            });
         }
         node.tagged_flow.tag.policy_data = node.policy_data_cache.clone();
     }