--- a/agent/src/flow_generator/protocol_logs/rpc/protobuf_rpc/krpc.rs
+++ b/agent/src/flow_generator/protocol_logs/rpc/protobuf_rpc/krpc.rs
@@ -184,7 +184,7 @@ pub struct KrpcLog {
 impl Default for KrpcLog {
     fn default() -> Self {
         Self {
-            previous_log_info: LruCache::new(100),
+            previous_log_info: LruCache::new(100.try_into().unwrap()),
             info: KrpcInfo::default(),
             perf_stats: None,
             parsed: false,